Move ColumnMetrics into ColumnUtils
