+++ packages/react-data-grid/src/Header.tsx
@@ -2,7 +2,7 @@ import React, { forwardRef, useRef, useState, useMemo, useImperativeHandle } fro
 import classNames from 'classnames';
 
 import HeaderRow from './HeaderRow';
-import { getColumnMetrics } from './ColumnMetrics';
+import { getColumnMetrics } from './utils/ColumnUtils';
 import { getScrollbarSize } from './utils';
 import { CalculatedColumn, HeaderRowData } from './common/types';
 import { GridProps } from './Grid';
