+++ packages/react-data-grid/src/ReactDataGrid.tsx
@@ -9,7 +9,7 @@ import React, {
 } from 'react';
 
 import Grid from './Grid';
-import { getColumnMetrics } from './ColumnMetrics';
+import { getColumnMetrics } from './utils/ColumnUtils';
 import EventBus from './EventBus';
 import { CellNavigationMode, EventTypes, UpdateActions, HeaderRowType, DEFINE_SORT } from './common/enums';
 import {
